+++ spi/src/main/java/hosh/spi/Values.java
@@ -23,8 +23,6 @@
  */
 package hosh.spi;
 
-import hosh.doc.Experimental;
-
 import java.io.PrintWriter;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -558,7 +556,6 @@ public class Values {
 		}
 	}
 
-	@Experimental(description = "new API")
 	public static class Comparators {
 
 		private Comparators() {
@@ -577,6 +574,10 @@ public class Values {
 
 		public static Comparator<Value> noneLast(Comparator<Value> comparator) {
 			return new NoneLastComparator(comparator);
+		}
+
+		public static Comparator<Value> noneFirst(Comparator<Value> comparator) {
+			return new NoneFirstComparator(comparator);
 		}
 
 		static class AlphaNumericStringComparator implements Comparator<String> {
@@ -632,6 +633,27 @@ public class Values {
 				}
 			}
 		}
+
+		static class NoneFirstComparator implements Comparator<Value> {
+
+			private final Comparator<Value> inner;
+
+			private NoneFirstComparator(Comparator<Value> inner) {
+				this.inner = inner;
+			}
+
+			@Override
+			public int compare(Value a, Value b) {
+				if (a instanceof None) {
+					return (b instanceof None) ? 0 : -1;
+				} else if (b instanceof None) {
+					return 1;
+				} else {
+					return inner.compare(a, b);
+				}
+			}
+		}
+
 
 	}
 }